Clean more and more tests
